+++ src/AudioOutput/WaveOutput.h
@@ -145,7 +145,7 @@ public:
 
     int write (int32_t* buffer, frame_t frames) override;
 
-    void setVolume(uint8_t) override;
+    void setVolume(float) override;
 
     void start () override;
     void stop () override;
