+++ src/AudioOutput/ebur128Output.cpp
@@ -153,7 +153,7 @@ int ebur128Output::write (int32_t* buffer, frame_t frames)
     THROW_RUNTIME_ERROR("ebur128: out of memory or something else");
 }
 
-void ebur128Output::setVolume(uint8_t)
+void ebur128Output::setVolume(float)
 {
     
 }
